requirements.txt
# --- Core Frameworks ---
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# --- LLM Providers ---
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# --- Backend & API ---
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1

# --- Data Science & Processing ---
pandas>=2.2.0
numpy>=1.26.0

# --- Evaluation & Monitoring ---
deepeval>=0.21.0
langsmith>=0.1.0